requirements.txt
openai
pinecone
pypdf
streamlit
python-dotenv
